requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
openai==1.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.18
pydantic==1.10.11
dateparser==1.1.8
requests==2.31.0
